Lint + minor refactor
