+++ benchmarks/facility_pyoframe.py
@@ -73,5 +73,6 @@ def main(Ns=[25, 50, 75, 100]):
                 io.write(f"{content}\n")
     return
 
+
 if __name__ == "__main__":
     main()
