+++ benchmarks/facility_pyopti.py
@@ -89,5 +89,6 @@ def main(Ns=[25, 50, 75, 100]):
             io.write(f"{content}\n")
     return
 
+
 if __name__ == "__main__":
     main()
